+++ packages/augur-sdk/src/state/getter/Users.ts
@@ -10,6 +10,12 @@ import {
   MarketCreatedLog,
 } from '../logs/types';
 import {
+  DisputeCrowdsourcerRedeemed,
+  MarketFinalized,
+  OrderEvent,
+  ProfitLossChanged,
+} from '../../event-handlers';
+import {
   Augur,
   numTicksToTickSize,
   convertOnChainAmountToDisplayAmount,
@@ -48,6 +54,27 @@ const getProfitLossParams = t.intersection([
     outcome: t.number,
   }),
 ]);
+
+export interface AccountTimeRangedStatsResult {
+  // Yea. The ProfitLossChanged event then
+  // Sum of unique entries (defined by market + outcome) with non-zero netPosition
+  positions: number;
+
+  // OrderEvent table for fill events (eventType == 3) where they are the orderCreator or orderFiller address
+  // if multiple fills in the same tx count as one trade then also counting just the unique tradeGroupId from those
+  numberOfTrades: number;
+
+  marketsCreated: number;
+
+  // Trades? uniq the market
+  marketsTraded: number;
+
+  // DisputeCrowdsourcerRedeemed where the payoutNumerators match the MarketFinalized winningPayoutNumerators
+  successfulDisputes: number;
+
+  // For getAccountTimeRangedStats.redeemedPositions use the InitialReporterRedeemed and DisputeCrowdsourcerRedeemed log?
+  redeemedPositions: number;
+}
 
 export interface MarketTradingPosition {
   timestamp: number;
@@ -106,12 +133,143 @@ export interface ProfitLossResult {
 }
 
 export class Users {
+  static getAccountTimeRangedStatsParams = t.intersection([
+  t.type({
+    universe: t.string,
+    account: t.string,
+  }),
+  t.partial({
+    endTime: t.number,
+    startTime: t.number,
+  })]);
+
   static getUserTradingPositionsParams = t.intersection([
     userTradingPositionsParams,
     sortOptions,
   ]);
   static getProfitLossParams = getProfitLossParams;
   static getProfitLossSummaryParams = getProfitLossSummaryParams;
+
+  @Getter("getAccountTimeRangedStatsParams")
+  static async getAccountTimeRangedStats(
+    augur: Augur,
+    db: DB,
+    params: t.TypeOf<typeof Users.getAccountTimeRangedStatsParams>
+  ): Promise<AccountTimeRangedStatsResult> {
+
+    // guards
+    if (!(await augur.contracts.augur.isKnownUniverse_(params.universe))) {
+      throw new Error('Unknown universe: ' + params.universe);
+    }
+
+    const startTime = params.startTime ? params.startTime : 0;
+    const endTime = params.endTime ? params.endTime : await augur.contracts.augur.getTimestamp_();
+
+    if (params.startTime > params.endTime) {
+      throw new Error("startTime must be less than or equal to endTime");
+    }
+
+    const baseRequest = {
+      universe: params.universe,
+      $and: [
+        { timestamp: { $gte: `0x${startTime.toString(16)}` } },
+        { timestamp: { $lte: `0x${endTime.toString(16)}` } },
+      ],
+    };
+
+    const marketsRequest = {
+      selector: Object.assign({
+        marketCreator: params.account,
+      }, baseRequest),
+    };
+
+    const initialReporterRequest = {
+      selector: Object.assign({
+        reporter: params.account,
+      }, baseRequest),
+    };
+
+    const disputeCrowdourcerRequest = {
+      selector: Object.assign({
+        disputeCrowdsourcerer: params.account,
+      }, baseRequest),
+    };
+
+    const profitLossChangedRequest = {
+      selector: Object.assign({
+        account: params.account,
+        netPosition: { $ne: 0 },
+      }, baseRequest),
+    };
+
+    const orderFilledRequest = {
+      selector: Object.assign({
+        orderCeator: params.account,
+        orderFiller: params.account,
+      }, baseRequest),
+    };
+
+    const compareArrays = (lhs: string[], rhs: string[]): number => {
+      let equal = 1;
+
+      lhs.forEach((item: string, index: number) => {
+        if (index >= rhs.length || item !== rhs[index]) {
+          equal = 0;
+        }
+      });
+
+      return equal;
+    };
+
+    const marketsCreatedLog = await db.findMarketCreatedLogs(marketsRequest);
+
+    const marketsCreated = marketsCreatedLog.length;
+    const initialReporterReedeemedLogs = await db.findInitialReporterRedeemedLogs(initialReporterRequest);
+    const disputeCrowdsourcerReedeemedLogs = await db.findDisputeCrowdsourcerRedeemedLogs(disputeCrowdourcerRequest);
+
+    const successfulDisputes = _.sum(await Promise.all((disputeCrowdsourcerReedeemedLogs as any as DisputeCrowdsourcerRedeemed[])
+      .map(async (log: DisputeCrowdsourcerRedeemed) => {
+        const marketFinalization = {
+          selector: Object.assign({
+            market: log.market,
+          }, baseRequest),
+        };
+        const markets = (await db.findMarketFinalizedLogs(marketFinalization)) as any as MarketFinalized[];
+
+        if (markets.length) {
+          return compareArrays(markets[0].winningPayoutNumerators, log.payoutNumerators);
+        }
+        else {
+          return 0;
+        }
+      })));
+
+
+    const redeemedPositions = initialReporterReedeemedLogs.length + successfulDisputes;
+
+    const orderFilledLogs = await db.findOrderFilledLogs(orderFilledRequest);
+    const numberOfTrades = _.uniqWith(orderFilledLogs as any as OrderEvent[], (a: OrderEvent, b: OrderEvent) => {
+      return a.tradeGroupId === b.tradeGroupId;
+    }).length;
+
+    const marketsTraded = _.uniqWith(orderFilledLogs as any as OrderEvent[], (a: OrderEvent, b: OrderEvent) => {
+      return a.market === b.market;
+    }).length;
+
+    const profitLossChangedLogs = await db.findProfitLossChangedLogs(params.account, profitLossChangedRequest);
+    const positions = _.uniqWith(profitLossChangedLogs as any as ProfitLossChanged[], (a: ProfitLossChanged, b: ProfitLossChanged) => {
+      return a.market === b.market && a.outcome === b.outcome;
+    }).length;
+
+    return {
+      positions,
+      numberOfTrades,
+      marketsCreated,
+      marketsTraded,
+      successfulDisputes,
+      redeemedPositions,
+    };
+  }
 
   @Getter('getUserTradingPositionsParams')
   static async getUserTradingPositions(
